refactor: move ConfigButtons into shared folder
