+++ apps/website/src/app/[locale]/admin/page.test.tsx
@@ -3,11 +3,11 @@ import { render, screen, cleanup } from "@testing-library/react";
 import Page from "./page";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
-vi.mock("@/components/LanguageToggleButton/LanguageToggleButton", () => ({
+vi.mock("@/components/ConfigButtons/LanguageToggleButton", () => ({
   default: () => <div data-testid="language-toggle-button">Language Toggle</div>,
 }));
 
-vi.mock("@/components/ThemeToggleButton/ThemeToggleButton", () => ({
+vi.mock("@/components/ConfigButtons/ThemeToggleButton", () => ({
   default: () => <div data-testid="theme-toggle-button">Theme Toggle</div>,
 }));
