+++ apps/website/src/app/[locale]/admin/page.tsx
@@ -4,8 +4,7 @@ import { Title } from "@flash/ui";
 import styles from "./AdminLogin.module.css";
 import CameraIcon from "@/components/Logo/Logo";
 import { useTranslations } from "next-intl";
-import LanguageToggleButton from "@/components/LanguageToggleButton/LanguageToggleButton";
-import ThemeToggleButton from "@/components/ThemeToggleButton/ThemeToggleButton";
+import { LanguageToggleButton, ThemeToggleButton } from "@/components/ConfigButtons";
 
 export default function Page() {
   const t = useTranslations("pages.adminLogin");
