+++ apps/website/src/app/[locale]/page.test.tsx
@@ -34,11 +34,11 @@ vi.mock("@/providers/JoinedEventsContext", () => ({
   useJoinedEvents: vi.fn(() => []),
 }));
 
-vi.mock("@/components/LanguageToggleButton/LanguageToggleButton", () => ({
+vi.mock("@/components/ConfigButtons/LanguageToggleButton", () => ({
   default: () => <div data-testid="language-toggle-button">Language Toggle</div>,
 }));
 
-vi.mock("@/components/ThemeToggleButton/ThemeToggleButton", () => ({
+vi.mock("@/components/ConfigButtons/ThemeToggleButton", () => ({
   default: () => <div data-testid="theme-toggle-button">Theme Toggle</div>,
 }));
