+++ apps/website/src/components/ConfigButtons/index.ts
@@ -1,0 +1,2 @@
+export { default as LanguageToggleButton } from "./LanguageToggleButton";
+export { default as ThemeToggleButton } from "./ThemeToggleButton";
