--- apps/website/src/components/LanguageToggleButton/LanguageToggleButton.module.css
@@ -1,44 +1,0 @@
-.button {
-  position: absolute;
-  top: 2rem;
-  right: 2rem;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  min-height: 2.5rem;
-  padding: 0.5rem 0.85rem;
-  border: 0.125rem solid var(--border-strong);
-  border-radius: var(--radius-full);
-  background-color: var(--bg-raised);
-  color: var(--text-default);
-  font-family: var(--font-family);
-  font-size: var(--font-size-small);
-  cursor: pointer;
-  transition:
-    background-color 0.2s ease,
-    border-color 0.2s ease,
-    transform 0.15s ease;
-
-  &:hover {
-    background-color: var(--bg-moderate-hoverT);
-  }
-
-  &:active {
-    background-color: var(--bg-moderate-pressedT);
-    transform: translateY(1px);
-  }
-
-  &:focus-visible {
-    outline: 2px solid var(--border-focus);
-    outline-offset: 2px;
-  }
-}
-
-@media (max-width: 768px) {
-  .button {
-    min-height: 2.75rem;
-    padding: 0.65rem 0.95rem;
-    font-size: var(--font-size-medium);
-  }
-}
